pyproject.toml
[tool.poetry]
name = "set_me_up"
version = "0.1.0"
description = ""
authors = ["andrewjunyoung"]

[tool.poetry.dependencies]
python = "^3.7"
wget = "^3.2"
gitpython = "^3.1.3"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
